Fixed '*# ' command operation.
Added 'hack' to allow the 'numparams' variable to be set (or unset)
before the netlist part of a file is parsed. This is based on code supplied by
Dietmar Warning, with  changes to allow it to work with the *# syntax, the
unset keyword and variable amounts of white space.

line_free(line,flag) is now a macro so that line is set to NULL when it is
freed. (src/frontend/inp.c) This is a safty mechnism.