parser/inpgmod.c, fix an almost bug, (level=...)
When 'level' was a valid model parameter,
  it was processed like any other parameter,
but additionally the
  ´if (strcmp(parm. "level"))'
invoked INPgetValue() a second time.

This didn't cause harm, merely because the next token after the
  level=number
almost necessarily is a string (the bginning parameter= of the next assignment)
thus not a parse-able number, and
thus the second INPgetValue didn't modify the 'line' pointer.

For clarity invoke the "level" skipping only if "level"
  is not a valid model parameter.