Incomplete and experimental change, for review only: do not merge. In particular only two of four types of plotting in ft_graf() (plotcurv.c) are implemented.
The purpose of this commit is plot large (more than MAXPOINTS) vectors
incrementally when plotting to a screen, returning to the graphics
systems' event loop after each burst of plotting.  That lets
UI interaction continue during plotting, regardless of vector size.

The data structure to hold incremental plotting context is in fteext.h.
The event loops in winmain.c and x11.c have been modified to support
plotting as a background activity with additional changes in windisp.c.
New functions are defined in graf.c.  In plotit.c, initial plotting
to the screen has been removed, so that all such plotting is now
started by handlers for WM_PAINT or expose events.  The main part is
in ft_graf() (plotcurv.c) which has been modified to plot vectors
incrementally, rather than all in a single call.